requirements.txt
streamlit==1.32.2
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
plotly==5.18.0
pandas==2.1.4
streamlit-option-menu==0.3.6
